requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./requirements.in
#
aiohttp==3.8.5
    # via web3
aiosignal==1.3.1
    # via aiohttp
anyio==3.7.1
    # via taskiq
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
base58==1.0.3
    # via
    #   py-cid
    #   py-multihash
bitarray==2.8.1
    # via eth-account
cached-property==1.5.2
    # via
    #   py-ecc
    #   py-evm
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   eth-ape
    #   silverback
commonmark==0.9.1
    # via rich
cryptography==41.0.3
    # via pyjwt
cytoolz==0.12.2
    # via eth-utils
dataclassy==0.11.1
    # via eip712
decorator==5.1.1
    # via ipython
deprecated==1.2.14
    # via pygithub
eip712==0.2.1
    # via eth-ape
eth-abi==4.1.0
    # via
    #   eip712
    #   eth-account
    #   eth-ape
    #   eth-tester
    #   web3
eth-account==0.8.0
    # via
    #   eip712
    #   eth-ape
    #   eth-tester
    #   web3
eth-ape @ git+https://github.com/ApeWorX/ape.git@refs/pull/1624/head
    # via
    #   -r ./requirements.in
    #   silverback
eth-bloom==2.0.0
    # via py-evm
eth-hash[pycryptodome,pysha3]==0.5.2
    # via
    #   eip712
    #   eth-bloom
    #   eth-tester
    #   eth-utils
    #   trie
    #   web3
eth-keyfile==0.6.1
    # via eth-account
eth-keys==0.4.0
    # via
    #   eth-account
    #   eth-keyfile
    #   eth-tester
    #   py-evm
eth-rlp==0.3.0
    # via eth-account
eth-tester[py-evm]==0.9.1b1
    # via web3
eth-typing==3.4.0
    # via
    #   eip712
    #   eth-abi
    #   eth-ape
    #   eth-keys
    #   eth-utils
    #   py-ecc
    #   py-evm
    #   web3
eth-utils==2.2.0
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-ape
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   eth-tester
    #   ethpm-types
    #   evm-trace
    #   py-ecc
    #   py-evm
    #   rlp
    #   trie
    #   web3
ethpm-types==0.5.4
    # via
    #   eth-ape
    #   evm-trace
evm-trace==0.1.0a22
    # via eth-ape
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
executing==1.2.0
    # via stack-data
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
hexbytes==0.3.1
    # via
    #   eip712
    #   eth-account
    #   eth-ape
    #   eth-rlp
    #   ethpm-types
    #   trie
    #   web3
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
ijson==3.2.3
    # via eth-ape
importlib-metadata==6.8.0
    # via
    #   eth-ape
    #   taskiq
iniconfig==2.0.0
    # via pytest
ipython==8.14.0
    # via eth-ape
jedi==0.19.0
    # via ipython
jsonschema==4.19.0
    # via web3
jsonschema-specifications==2023.7.1
    # via jsonschema
lazyasd==0.1.4
    # via eth-ape
lru-dict==1.2.0
    # via
    #   py-evm
    #   web3
matplotlib-inline==0.1.6
    # via ipython
morphys==1.0
    # via
    #   py-cid
    #   py-multibase
    #   py-multicodec
    #   py-multihash
msgspec==0.18.1
    # via evm-trace
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   py-ecc
    #   py-evm
numpy==1.25.2
    # via pandas
packaging==23.1
    # via
    #   eth-ape
    #   pytest
pandas==1.5.3
    # via eth-ape
parsimonious==0.9.0
    # via eth-abi
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.2.0
    # via
    #   eth-ape
    #   pytest
prometheus-client==0.17.1
    # via taskiq
prompt-toolkit==3.0.39
    # via ipython
protobuf==4.24.1
    # via web3
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py-cid==0.3.0
    # via ethpm-types
py-ecc==6.0.0
    # via py-evm
py-evm==0.7.0a4
    # via
    #   eth-tester
    #   evm-trace
py-geth==3.13.0
    # via
    #   eth-ape
    #   web3
py-multibase==1.0.3
    # via py-cid
py-multicodec==0.2.1
    # via py-cid
py-multihash==0.2.3
    # via py-cid
pycparser==2.21
    # via cffi
pycron==3.0.0
    # via taskiq
pycryptodome==3.18.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==1.10.12
    # via
    #   eth-ape
    #   ethpm-types
    #   evm-trace
    #   taskiq
pyethash==0.1.27
    # via py-evm
pygithub==1.59.1
    # via eth-ape
pygments==2.16.1
    # via
    #   ipython
    #   rich
pyjwt[crypto]==2.8.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyrate-limiter==2.10.0
    # via requests-ratelimiter
pytest==7.4.0
    # via eth-ape
python-baseconv==1.2.2
    # via py-multibase
python-dateutil==2.8.2
    # via
    #   eth-ape
    #   pandas
pytz==2023.3
    # via pandas
pyunormalize==15.0.0
    # via web3
pyyaml==6.0.1
    # via eth-ape
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.8.8
    # via parsimonious
requests==2.31.0
    # via
    #   -r ./requirements.in
    #   eth-ape
    #   ethpm-types
    #   pygithub
    #   requests-ratelimiter
    #   web3
requests-ratelimiter==0.4.1
    # via -r ./requirements.in
rich==12.6.0
    # via eth-ape
rlp==3.0.0
    # via
    #   eth-account
    #   eth-rlp
    #   eth-tester
    #   py-evm
    #   trie
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
safe-pysha3==1.0.4
    # via eth-hash
semantic-version==2.10.0
    # via
    #   eth-tester
    #   py-geth
silverback==0.1.0b11
    # via -r ./requirements.in
six==1.16.0
    # via
    #   asttokens
    #   py-multibase
    #   py-multicodec
    #   py-multihash
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sortedcontainers==2.4.0
    # via trie
sqlalchemy==2.0.20
    # via eth-ape
stack-data==0.6.2
    # via ipython
taskiq[metrics]==0.6.0
    # via silverback
taskiq-dependencies==1.3.1
    # via taskiq
tomli==2.0.1
    # via pytest
toolz==0.12.0
    # via cytoolz
tqdm==4.66.1
    # via eth-ape
traitlets==5.9.0
    # via
    #   eth-ape
    #   ipython
    #   matplotlib-inline
trie==2.1.1
    # via py-evm
typing-extensions==4.7.1
    # via
    #   pydantic
    #   sqlalchemy
    #   taskiq
    #   web3
urllib3==2.0.4
    # via requests
varint==1.0.2
    # via
    #   py-multicodec
    #   py-multihash
watchdog==3.0.0
    # via eth-ape
wcwidth==0.2.6
    # via prompt-toolkit
web3[tester]==6.9.0
    # via eth-ape
websockets==11.0.3
    # via web3
wrapt==1.15.0
    # via deprecated
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via importlib-metadata
